backend/requirements.txt
asgiref==3.8.1
Django==5.0.7
django-allauth==65.0.2
django-cors-headers==4.5.0
django-rest-auth==0.9.5
djangorestframework==3.15.2
six==1.16.0
sqlparse==0.5.1
